// app/src/interfaces/empleadoInterfaces.tsx
export interface UseEmpleado {
    id_empleado: number;
    nombre: string;
    curp: string;
    rfc: string;
    nss: string;
    fecha_nacimiento: string;
    direccion: string;
    telefono: string;
    correo: string;
    cargo: string;
    area: string;
    fecha_ingreso: string;
    cuenta_bancaria: string;
    contacto_exta: string;
    estado_civil: string;
    estudios: string;
    foto_perfil: string;
    ine_anversa: string;
    ine_reverso: string;
    comprobante_domicilio: string;
    constancia_fiscal: string;
    acta_nacimiento: string;
}

export type FormDataEmpleado = UseEmpleado;